refactor(playground): inline usePlaygroundServers into Playground
Used in one place; a separate 80-line file was scaffolding.
